@@ -1,15 +1,19 @@
 plugins {
     id 'java'
 }
-
+group 'com.example'
 version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
+sourceCompatibility = 11
 repositories {
     mavenCentral()
 }
-
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
+    testCompile('org.junit.jupiter:junit-jupiter-api:5.3.2')
+    testRuntime('org.junit.jupiter:junit-jupiter-engine:5.3.2')
+}
+test {
+
+    13
+
+    useJUnitPlatform()
 }
